Re-organises project packages
