+++ src/main/kotlin/io/dotanuki/arw/Main.kt
@@ -1,8 +1,8 @@
 package io.dotanuki.arw
 
 import com.github.ajalt.clikt.core.subcommands
-import io.dotanuki.arw.cli.ARWCommand
-import io.dotanuki.arw.overview.OverviewCommand
+import io.dotanuki.arw.core.infrastructure.cli.ARWCommand
+import io.dotanuki.arw.features.overview.OverviewCommand
 
 fun main(args: Array<String>) {
     ARWCommand()
