+++ src/test/kotlin/io/dotanuki/arw/helpers/ErrorAwareAssertion.kt
@@ -1,7 +1,7 @@
 package io.dotanuki.arw.helpers
 
 import arrow.core.raise.recover
-import io.dotanuki.arw.shared.errors.ErrorAware
+import io.dotanuki.arw.core.domain.errors.ErrorAware
 
 fun errorAwareTest(assertion: ErrorAware.() -> Unit) =
     recover(
